requirements.txt
numpy
pandas
mysql-connector-python
scikit-learn
setuptools